# src/main/resources/openapi/specs.yaml
openapi: 3.1.0
info:
  title: API
  version: 1.0.0
paths:
  /clients:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - first_name
                - last_name
                - email
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                  format: email
                countryOfResidence:
                  type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
          description: 'tmp description'
  /clients/{id}/documents:
    post:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
          description: 'tmp description'
  /search:
    get:
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
          description: 'tmp description'
components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        countryOfResidence:
          type: string
    Document:
      type: object
      properties:
        id:
          type: string
        client_id:
          type: string
        title:
          type: string
        content:
          type: string
        created_at:
          type: string
          format: date-time